# perf: improve performance of resultSet.getObject