There are two changes:
1) Fast-path for `long` -> BigDecimal. For small values, database can send binary `long`, so there is no need to convert the received long back to string (see type=BIGINT below)
2) "Cached getSQLType" speeds up getObject (see type=bigdecimal, getter=object below).

Before:
```
Benchmark                 (getter)  (ncols)  (nrows)     (type)  Mode  Cnt      Score    Error  Units
ProcessResultSet.execFetch    BEST        5     1000         INT  avgt    5   599,098 ± 11,030  us/op
ProcessResultSet.execFetch    BEST        5     1000  BIGDECIMAL  avgt    5  2392,760 ± 26,083  us/op

ProcessResultSet.execFetch    BEST        5     1000      BIGINT  avgt    5  3141,686 ± 92,846  us/op
ProcessResultSet.execFetch  OBJECT        5     1000  BIGDECIMAL  avgt    5  3400,252 ± 61,252  us/op
```

After:
```
Benchmark                 (getter)  (ncols)  (nrows)      (type)  Mode  Cnt     Score    Error  Units
ProcessResultSet.execFetch    BEST        5     1000         INT  avgt    5   598,402 ± 10,890  us/op
ProcessResultSet.execFetch    BEST        5     1000  BIGDECIMAL  avgt    5  2370,942 ± 31,716  us/op

ProcessResultSet.execFetch    BEST        5     1000      BIGINT  avgt    5   719,552 ±  9,449  us/op
ProcessResultSet.execFetch  OBJECT        5     1000  BIGDECIMAL  avgt    5  2450,741 ± 52,455  us/op
```