@davecramer 2015-11-24
Yes, this has already been fixed in
https://github.com/pgjdbc/pgjdbc/pull/426

Thanks for the report

Dave Cramer

On 24 November 2015 at 02:59, Mario Ivankovits <notifications@github.com>
wrote:

> Copied over my comment from the forked repository commit.
> There is an interesting, but huge problem with this commit.
>
> executing the following simple code snipped reveals that after 5
> iterations the database switch to binary transfer and then the decoding at
> line 2547 is using a scale of "-1" (the getBigDecimal default) instead of
> "0".
> PostgresSQL Version 9.4.4
>
>     Connection conn = DriverManager.getConnection(.....);
>     for (int i = 0; i<10; i++)
>     {
>         PreparedStatement stmt = conn.prepareStatement("SELECT count(*) FROM pg_tables");
>         ResultSet rs = stmt.executeQuery();
>         rs.next();
>         Number nb = rs.getBigDecimal(1);
>         System.err.println("#" + i + "=" + nb);
>         stmt.close();
>     }
>
> Result:
>
> #0=250
> #1 <https://github.com/pgjdbc/pgjdbc/pull/1>=250
> #2 <https://github.com/pgjdbc/pgjdbc/issues/2>=250
> #3 <https://github.com/pgjdbc/pgjdbc/issues/3>=250
> #4 <https://github.com/pgjdbc/pgjdbc/pull/4>=250
> #5 <https://github.com/pgjdbc/pgjdbc/pull/5>=2.50E+3
> #6 <https://github.com/pgjdbc/pgjdbc/issues/6>=2.50E+3
> #7 <https://github.com/pgjdbc/pgjdbc/pull/7>=2.50E+3
> #8 <https://github.com/pgjdbc/pgjdbc/pull/8>=2.50E+3
> #9 <https://github.com/pgjdbc/pgjdbc/pull/9>=2.50E+3
>
> As a side note: It might be worth to check if internalGetObject already
> returned a BigDecimal and so you can return that instead of converting it
> to a string and then back to a bigdecimal using the trimMoney stuff.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pgjdbc/pgjdbc/pull/411#issuecomment-159189252>.
>